[LOH] Coherence format int algo

+++ PLD-Agile/src/modele/algo/TSP.java
@@ -16,7 +16,7 @@ public interface TSP {
 	 * @param cout : cout[i][j] = duree pour aller de i a j, avec 0 <= i < nbSommets et 0 <= j < nbSommets
 	 * @param duree : duree[i] = duree pour visiter le sommet i, avec 0 <= i < nbSommets
 	 */
-	public Integer[] chercheSolution(int tpsLimite, int nbSommets, float[][] cout, int[] duree, Long[][] horaires);
+	public Integer[] chercheSolution(int tpsLimite, int nbSommets, float[][] cout, int[] duree, int[][] horaires);
 	
 	/**
 	 * @param i
